use DefaultTimeout with params
