+++ aws/internal/service/storagegateway/waiter/status.go
@@ -13,21 +13,15 @@ import (
 )
 
 const (
-	StorageGatewayGatewayStatusConnected        = "GatewayConnected"
-	StoredIscsiVolumeStatusNotFound             = "NotFound"
-	StoredIscsiVolumeStatusUnknown              = "Unknown"
-	NfsFileShareStatusNotFound                  = "NotFound"
-	NfsFileShareStatusUnknown                   = "Unknown"
-	SmbFileShareStatusNotFound                  = "NotFound"
-	SmbFileShareStatusUnknown                   = "Unknown"
-	FsxFileSystemStatusNotFound                 = "NotFound"
-	FsxFileSystemStatusUnknown                  = "Unknown"
-	FsxFileSystemAssociationStatusAvailable     = "AVAILABLE"
-	FsxFileSystemAssociationStatusCreating      = "CREATING"
-	FsxFileSystemAssociationStatusDeleting      = "DELETING"
-	FsxFileSystemAssociationStatusForceDeleting = "FORCE_DELETING"
-	FsxFileSystemAssociationStatusUpdating      = "UPDATING"
-	FsxFileSystemAssociationStatusError         = "ERROR"
+	StorageGatewayGatewayStatusConnected = "GatewayConnected"
+	StoredIscsiVolumeStatusNotFound      = "NotFound"
+	StoredIscsiVolumeStatusUnknown       = "Unknown"
+	NfsFileShareStatusNotFound           = "NotFound"
+	NfsFileShareStatusUnknown            = "Unknown"
+	SmbFileShareStatusNotFound           = "NotFound"
+	SmbFileShareStatusUnknown            = "Unknown"
+	FsxFileSystemStatusNotFound          = "NotFound"
+	FsxFileSystemStatusUnknown           = "Unknown"
 )
 
 func StorageGatewayGatewayStatus(conn *storagegateway.StorageGateway, gatewayARN string) resource.StateRefreshFunc {
@@ -154,20 +148,4 @@ func FsxFileSystemStatus(conn *storagegateway.StorageGateway, fileSystemArn stri
 
 		return output, aws.StringValue(output.FileSystemAssociationStatus), nil
 	}
-}
-
-func FsxFileSystemStatusAvailableStatusPending() []string {
-	return []string{FsxFileSystemAssociationStatusCreating, FsxFileSystemAssociationStatusUpdating}
-}
-
-func FsxFileSystemStatusAvailableStatusTarget() []string {
-	return []string{FsxFileSystemAssociationStatusAvailable}
-}
-
-func FsxFileSystemStatusDeletedStatusPending() []string {
-	return []string{FsxFileSystemAssociationStatusAvailable, FsxFileSystemAssociationStatusDeleting, FsxFileSystemAssociationStatusForceDeleting}
-}
-
-func FsxFileSystemStatusDeletedStatusTarget() []string {
-	return []string{}
 }
