+++ aws/resource_aws_storagegateway_fsx_associate_file_system.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"log"
 	"regexp"
-	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/storagegateway"
@@ -29,9 +28,9 @@ func resourceAwsStorageGatewayFsxAssociateFileSystem() *schema.Resource {
 			State: schema.ImportStatePassthrough,
 		},
 		Timeouts: &schema.ResourceTimeout{
-			Create: schema.DefaultTimeout(3 * time.Minute),
-			Update: schema.DefaultTimeout(3 * time.Minute),
-			Delete: schema.DefaultTimeout(3 * time.Minute),
+			Create: schema.DefaultTimeout(tfstoragegateway.AssociateFileSystemCreateTimeout),
+			Update: schema.DefaultTimeout(tfstoragegateway.AssociateFileSystemUpdateTimeout),
+			Delete: schema.DefaultTimeout(tfstoragegateway.AssociateFileSystemDeleteTimeout),
 		},
 
 		Schema: map[string]*schema.Schema{
